04_mlflow/requirements.txt
# MLflow Tracking Server Requirements

huggingface-hub==0.20.3
requests==2.31.0
mlflow==2.10.0
psycopg2-binary==2.9.9
boto3==1.34.34
python-dotenv==1.0.1
